[lldb-dap] Do not expose internal breakpoints to DAP clients
This commit fixes the following problem seen on Android:
When the process stops due to a user breakpoint, sometimes a second breakpoint is hit at the same time, and the corresponding thread is marked as stopped for reason `jit-debug-register`. This second breakpoint is used by lldb internally, to track code generated by the JIT compiler (JVM). Instead of displaying the code corresponding to the user breakpoint, the debugger displays the disassembled code corresponding to this internal breakpoint.